% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaTranslation.R
\name{aaTranslation}
\alias{aaTranslation}
\title{AminoAcid translation of the DNAStringSet object}
\usage{
aaTranslation(subalign, DNAStr)
}
\arguments{
\item{subalign}{the subalignment of DNA sequences that will be used for translation as DNAStringSet object}

\item{DNAStr}{DNAStringSet object that has the multiple sequence alignment (it is used for just names of the sequences)}
}
\value{
AAStringSet object of translated and aligned AminoAcid sequences if there is no error.
if there is error, it will show a message and return FALSE. It should be checked in the main script
}
\description{
AminoAcid translation of the DNAStringSet object
}
