% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Gets DNAStringSet aligns it}
\usage{
align(mySequences, orfName, path)
}
\arguments{
\item{mySequences}{DNAStringSet with all sequences (including the ORF to be searched) in it}

\item{orfName}{ORF identifier name that will be used for file writing}

\item{path}{path for files to be written in}
}
\value{
aligned DNAStringset, start and stop positions of orfName on DNAStringSet and AA alignment of the orfName sequence with other species
}
\description{
Gets DNAStringSet aligns it
}
