% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignAnalyze.R
\name{alignAnalyze}
\alias{alignAnalyze}
\title{This will be the function for alignment-analysis output}
\usage{
alignAnalyze(filename, orfName, annotated = T, outputDirectory, ...)
}
\arguments{
\item{filename}{alignment file name or directory name}

\item{orfName}{identifier for ORF. SGD name if annotated}

\item{annotated}{boolean shows if the given sequence is annotated or not}

\item{outputDirectory}{directory to save alignment output}

\item{...}{placeholder for sequence file names}
}
\description{
This will be the function for alignment-analysis output
}
