% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzePairwise.R
\name{analyzePairwise}
\alias{analyzePairwise}
\title{Analyze pairwise alignment
this is helper function for analyze script. Creates a df for every overlapping orf}
\usage{
analyzePairwise(typeName, outputDirectory, orfName, startAA, startDNA,
  stopAA, subalign, aa, df, identifier = "best")
}
\arguments{
\item{typeName}{species name}

\item{outputDirectory}{folder path for reading input files}

\item{orfName}{identifier for ORF. SGD name if annotated}

\item{startAA}{aminoacid start codons calculated by \code{startCodon}}

\item{startDNA}{dna start codons calculated by \code{startCodon}}

\item{stopAA}{aa stop codons calculated by \code{stopCodon}}

\item{subalign}{dna level subalignment}

\item{aa}{aminoacid translation using subalignment}

\item{df}{df for reference orf values}

\item{identifier}{identifier for the overlapping orf name. it will be used for file names as well.}
}
\value{
dataFrame object
}
\description{
Analyze pairwise alignment
this is helper function for analyze script. Creates a df for every overlapping orf
}
