% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_identity}
\alias{calc_identity}
\title{calculate identity}
\usage{
calc_identity(aln, len)
}
\arguments{
\item{aln}{aligned sequences}

\item{len}{length of aligned sequence}
}
\value{
numeric vector of identical aminoacids between first seq in the aln object and the rest
}
\description{
calculate identity
}
