% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIdentity.R
\name{calcIdentity}
\alias{calcIdentity}
\title{Calculate %ID between sequences of alignments
Only %ID between Scer and other sequences are calculated
Gaps are treated as difference if one sequence has gap and other does not. If both sequence have gap in the same position it will not be count as similarity}
\usage{
calcIdentity(alignment)
}
\arguments{
\item{alignment}{DNAStringSet or AAStringSet object of alignments}
}
\value{
nx1 matrix with with %IDs
}
\description{
Calculate %ID between sequences of alignments
Only %ID between Scer and other sequences are calculated
Gaps are treated as difference if one sequence has gap and other does not. If both sequence have gap in the same position it will not be count as similarity
}
