% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAlignment.R
\name{checkAlignment}
\alias{checkAlignment}
\title{Check if all codons are same}
\usage{
checkAlignment(alignedLetter)
}
\arguments{
\item{alignedLetter}{list of AAString or DNAString instances, possibly result of startCodon or stopCodon functions}
}
\value{
TRUE if all same , FALSE otherwise
}
\description{
Check if all codons are same
}
