% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCodon.R
\name{checkCodon}
\alias{checkCodon}
\title{Check aminoacid character}
\usage{
checkCodon(aa_seq, checkCharacter)
}
\arguments{
\item{aa_seq}{Aminoacid sequence as character string. Change it by as.character if using AAString}

\item{checkCharacter}{the single letter aminoacid code for the searched character. 'M' for start and 'X' for stop}
}
\value{
the first occurence of the character.
}
\description{
Check aminoacid character
}
