% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBestOverlap.R
\name{checkFrame}
\alias{checkFrame}
\title{this function finds correct frame sequences of intersection and union of two Iranges objects}
\usage{
checkFrame(dna, map, unionStart, unionEnd, intersectStart, intersectEnd,
  start)
}
\arguments{
\item{dna}{ungapped sequence}

\item{map}{mapped list of original sequence-->gapped alignment created by {map_alignment_sequence}}

\item{unionStart}{union range start position}

\item{unionEnd}{union range last position}

\item{intersectStart}{intersection range start position}

\item{intersectEnd}{intersection range end position}

\item{start}{start of ORF of interest, to compare with unionStart and intersectStart values}
}
\value{
list of correct frame intersection and union dna sequences taken as subsequence of {dna}
}
\description{
this function finds correct frame sequences of intersection and union of two Iranges objects
}
