% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkORF.R
\name{checkORF}
\alias{checkORF}
\title{Check Orf}
\usage{
checkORF(aa_seq)
}
\arguments{
\item{aa_seq}{Aminoacid sequence as character string. Change it by as.character if using AAString}
}
\value{
if there is an ORF then return TRUE
If there is no ORF check what is missing an M or X
If M is missing return 'M'  --- if X is missing return 'X' --- if both are missing return 'BOTH', if there is not orf but X ---- M sequence, it will return 'Neither'
}
\description{
Checks if Amino Acid sequence has M --- X or I --- X ORF by using regular expressions. 
This can be changed to only check M --- X ORF, or for any other start codons
}
