% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLongerOverlap.R
\name{compareMandX}
\alias{compareMandX}
\title{this compares if there is X ---- M sequence, which one has larger overlap with Scer orf by checking the length between scer start to X and M to end of scer}
\usage{
compareMandX(aa_seq)
}
\arguments{
\item{aa_seq}{aminoacid sequence as AAString}
}
\value{
'M' or 'X'
}
\description{
this compares if there is X ---- M sequence, which one has larger overlap with Scer orf by checking the length between scer start to X and M to end of scer
}
