% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFrameShifting.R
\name{countFrameShifting}
\alias{countFrameShifting}
\title{Count frameshifts with MACSE}
\usage{
countFrameShifting(aaSeq)
}
\arguments{
\item{aaSeq}{macse aligned AA sequence}
}
\value{
number of '!' characters since MACSE uses them to introduce frameshifting indels to the alignments
}
\description{
Count frameshifts with MACSE
}
