% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBestOverlap.R
\name{findBestOverlap}
\alias{findBestOverlap}
\title{Finds best overlapping ORF in terms of the number of identical aminoacids}
\usage{
findBestOverlap(DNAStr, j, start, stop, ygeneSeq, types, k = NULL)
}
\arguments{
\item{DNAStr}{DNA alignment of syntenic block. This is the full alignment not subalignment}

\item{j}{this is the iterator for the ID of species of interest on DNAStr and types vector}

\item{start}{start position of ATG of proto-gene on DNAStr}

\item{stop}{last nucleotide position of proto-gene on DNAStr}

\item{ygeneSeq}{proto-gene sequence without gaps. This is used for determining the frame of proto-gene with respect to the aligned other species ORF}

\item{types}{this is used for names. by default c('Spar','Smik','Skud','Sbay')}
}
\value{
list of three elements

dna = dna alignment of jth species orf with proto-gene, aa = aa alignment with full jth ORF, aaOverlap = overlapping regions of both ORFs
}
\description{
Finds best overlapping ORF in terms of the number of identical aminoacids
}
