% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBoth.R
\name{findBoth}
\alias{findBoth}
\title{Search for both start and stop codons to get the overlapping orf}
\usage{
findBoth(DNAStr, speciesIndex, start, stop, orfName)
}
\arguments{
\item{DNAStr}{aligned DNAStringSet object}

\item{speciesIndex}{the index for which sequence in the DNAStringSet object to be searched for ATG and stop codons}

\item{start}{the start position of the S. Cerevisiae ORF over the alignment object}

\item{stop}{stop position of the S. Cerevisiae ORF over the alignment object. It is used for when the Start codon not found}

\item{orfName}{the corresponding orf name. Used for a clear warning message}
}
\value{
The list of boolean value for which will be used to determine if another subalignment and AA translation will be written and new start and stop positions
}
\description{
Search for both start and stop codons to get the overlapping orf
}
