% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCodon.R
\name{findCodon}
\alias{findCodon}
\title{find Codon
Searches for start or stop codons
For start codon search the input sequence should end right before the position where it alignes with start codon of proto-gene ORF 
For stop codon search the input sequence should start from the position where it alignes with start codon of proto-gene ORF so for actual position on the DNA alignment (that start - 1 ) should be added the returned value}
\usage{
findCodon(sequence, codonSet)
}
\arguments{
\item{sequence}{DNAString object}

\item{codonSet}{('ATG') for start codon or ('TAA', 'TAG', 'TGA') for stop codon search}
}
\value{
index of the A for start codon or index of last letter of stop codon for stop codon search
}
\description{
find Codon
Searches for start or stop codons
For start codon search the input sequence should end right before the position where it alignes with start codon of proto-gene ORF 
For stop codon search the input sequence should start from the position where it alignes with start codon of proto-gene ORF so for actual position on the DNA alignment (that start - 1 ) should be added the returned value
}
\seealso{
\code{\link{findBoth}} for usage
}
