% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHomolog.R
\name{findHomolog}
\alias{findHomolog}
\title{This function compares all overlapping ORFs with the Scer ORF and saves best ORF in terms of number of AAs (highest number of identical aminoacids)}
\usage{
findHomolog(DNAStr, aa_alignment, start, stop, ygeneSeq, types, path,
  orfName)
}
\arguments{
\item{DNAStr}{aligned DNAStringSet with all sequences (including the ORF to be searched) in it, this is the full syntenic block not subalignment}

\item{aa_alignment}{Amino acid alignment over the subalignment}

\item{start}{start codon position of ygeneSeq in DNAStr}

\item{stop}{last nucleotide position of ygeneSeq in DNAStr}

\item{ygeneSeq}{sequence of interest}

\item{types}{this is a vector which contains the species names in the alignment for writing correct output files}

\item{path}{path for files to be written in}

\item{orfName}{ORF identifier name that will be used for file writing}
}
\value{
nothing returns at the moment, all files are written in path with 3 files for each species. pairwise nucleotide alignment, pairwise AA alignment and pairwise AA alignment that shows only overlap of two ORFs
}
\description{
This function compares all overlapping ORFs with the Scer ORF and saves best ORF in terms of number of AAs (highest number of identical aminoacids)
}
