% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findI.R
\name{findI}
\alias{findI}
\title{Find all start codons}
\usage{
findI(aaSequence)
}
\arguments{
\item{aaSequence}{AAString instance}
}
\value{
coordinates of all I if there is any. -1 if none exists
}
\description{
Find all start codons
}
