% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findM.R
\name{findM}
\alias{findM}
\title{Find all start codons}
\usage{
findM(aaSequence)
}
\arguments{
\item{aaSequence}{AAString instance}
}
\value{
coordinates of all M if there is any. -1 if none exists
}
\description{
Find all start codons
}
