% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findORF.R
\name{findORF}
\alias{findORF}
\title{Find orf start and stop coordinates}
\usage{
findORF(aa_seq, findI = F)
}
\arguments{
\item{AAString}{instance or character vector}
}
\value{
coordinates of M---X or I---X. It calcualtes longest orf by considering all M to X pairs and first finds all non-overlapping ORFs and then returns the longest one.
}
\description{
Find orf start and stop coordinates
}
