% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlappingOrfs.R
\name{findOverlappingOrfs}
\alias{findOverlappingOrfs}
\title{Finds all overlapping ORFs}
\usage{
findOverlappingOrfs(dna, range)
}
\arguments{
\item{dna}{DNA sequence that will be searched for finding ORFs}

\item{range}{the range of interest. Only ORF ranges overlapping with this range will be considered}
}
\value{
IRanges object with start and end positions determined using \code{dna} object
}
\description{
Finds all overlapping ORFs
}
