% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlappingOrfsPar.R
\name{findOverlappingOrfsPar}
\alias{findOverlappingOrfsPar}
\title{Finds all overlapping ORFs}
\usage{
findOverlappingOrfsPar(dna, range)
}
\arguments{
\item{dna}{DNA sequence that will be searched for finding ORFs}

\item{range}{the range of interest. Only ORF ranges overlapping with this range will be considered}
}
\value{
IRanges object with start and end positions determined using \code{dna} object
}
\description{
Finds all overlapping ORFs
}
