% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findX.R
\name{findX}
\alias{findX}
\title{Find all stop codons}
\usage{
findX(aaSequence)
}
\arguments{
\item{aaSequence}{AAString instance}
}
\value{
coordinates of all X if there is any. -1 if none exists
}
\description{
Find all stop codons
}
