% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findYgeneSeq.R
\name{findYGeneSeq}
\alias{findYGeneSeq}
\title{Finds annotated ORF sequence from annotated sequence file downloaded from SGD}
\usage{
findYGeneSeq(geneName, outputPath = NULL)
}
\arguments{
\item{geneName}{annotated orf Y----- name}

\item{outputPath}{path if orfName sequence will be asked to written as separate file}
}
\value{
sequence of given orf name as DNAString instance
}
\description{
Finds annotated ORF sequence from annotated sequence file downloaded from SGD
}
