% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlappingOrfsPar.R
\name{irangesCombine}
\alias{irangesCombine}
\title{Combination function for IRanges to use in foreach}
\usage{
irangesCombine(r1, r2)
}
\arguments{
\item{r1}{iranges instance}

\item{r2}{iranges instance}
}
\value{
combined iranges object
}
\description{
Combination function for IRanges to use in foreach
}
