% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macseAdjustment.R
\name{macseAdjustment}
\alias{macseAdjustment}
\title{Correct macse
Changes some macse spesific characters for the general purposes of this package
since macse reorder the sequences after alignment, I use previously used alignment file to change the order back to scer,spar,smik,skud,sbay
changes '!' character to '-' and '*' to 'X'}
\usage{
macseAdjustment(macseAlignment, seqSet)
}
\arguments{
\item{macseAlignment}{macse aligned AAStringSet object}

\item{seqSet}{previously used DNAStringSet or AAStringSet object of the same sequences}
}
\value{
corrected AAStringSet object
}
\description{
Correct macse
Changes some macse spesific characters for the general purposes of this package
since macse reorder the sequences after alignment, I use previously used alignment file to change the order back to scer,spar,smik,skud,sbay
changes '!' character to '-' and '*' to 'X'
}
