% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{map_alignment_sequence}
\alias{map_alignment_sequence}
\title{map two sequences}
\usage{
map_alignment_sequence(aln, seq)
}
\arguments{
\item{aln}{gapped sequence}

\item{seq}{ungapped sequence}
}
\value{
a position map from seq->aln
}
\description{
map two sequences
}
