% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readFiles}
\alias{readFiles}
\title{Read syntetic block sequence files}
\usage{
readFiles(path)
}
\arguments{
\item{path}{folder path that has sequence fasta files in it
every sequence file needs to contain its species name . (i.e. S. Cerevisiae file should contain 'scer', proto-gene sequence should contain 'smorf' or 'sequence')
also assumes there is only 1 file for every species}
}
\value{
DNAStringSet object with the sequences in it 
order of DNAStringSet : Scer(1) -> Spar(2) -> Smik(3) -> Skud(4) --> Sbay(5) --> Smorf(6) 
if one of them does not exist, shows warning 

TODO check for multiple files having the search string in them
}
\description{
Read syntetic block sequence files
}
