% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regSeq.R
\name{regSeq}
\alias{regSeq}
\title{Add gaps to sequence
Search sequence with any number of gaps in between using regular expressions
adds '[-]*' in between every letter. This helps to seacrh orf sequence on the aligned sequence since there is no way to know where the gaps are}
\usage{
regSeq(seq)
}
\arguments{
\item{seq}{character string}
}
\value{
character string
}
\description{
Add gaps to sequence
Search sequence with any number of gaps in between using regular expressions
adds '[-]*' in between every letter. This helps to seacrh orf sequence on the aligned sequence since there is no way to know where the gaps are
}
