% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scerAnnotatedSequences.R
\docType{data}
\name{scerAnnotatedSequences}
\alias{scerAnnotatedSequences}
\title{Saccharomyces Cerevisiae annotated sequence data}
\format{An object of class \code{"DnaStringSet"}}
\source{
\href{https://www.yeastgenome.org/downloads}{SGD}
}
\usage{
data(scerAnnotatedSequences)
}
\description{
Saccharomyces Cerevisiae annotated sequence data
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
