% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesVector.R
\name{speciesVector}
\alias{speciesVector}
\title{This function check a set of strings to create a species vector}
\usage{
speciesVector(input)
}
\arguments{
\item{input}{a list of strings}
}
\value{
a vector that represents whether following species exists in this order. Scer,Spar,Smik,Skud,Sbay,Sarb
}
\description{
This function check a set of strings to create a species vector
}
