% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startCodon.R
\name{startCodon}
\alias{startCodon}
\title{Start Codon alignment}
\usage{
startCodon(alignment)
}
\arguments{
\item{alignment}{DNAStringSet or AAStringSet object of alignment}
}
\value{
list of first 3 nucleotide if DNAStringSet or first aminoacid if AAStringSet
}
\description{
Start Codon alignment
}
