% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopCodon.R
\name{stopCodon}
\alias{stopCodon}
\title{Stop codon alignment}
\usage{
stopCodon(alignment)
}
\arguments{
\item{DNAStringSet}{or AAStringSet object of alignment}
}
\value{
list of last 3 nucleotide if DNAStringSet or last aminoacid if AAStringSet
}
\description{
Stop codon alignment
}
