% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subalignment.R
\name{subalignment}
\alias{subalignment}
\title{Export subalingment of the region between start and stop Scer orf}
\usage{
subalignment(start, stop, DNAStr)
}
\arguments{
\item{start}{the start position of proto-gene on the DNAStr object}

\item{stop}{the stop position of the proto-gene on the DNAStr object (the position of last nucleotide of the ORF)}

\item{DNAStr}{aligned DNAStringSet object}
}
\value{
list of the subalignment as DNAStringSet object between the positions (start,stop) and identity percentage w.r.t. first sequence (SCER)
}
\description{
Export subalingment of the region between start and stop Scer orf
}
