% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnWoGaps.R
\name{turnWoGaps}
\alias{turnWoGaps}
\title{Remove gaps (\'-\')}
\usage{
turnWoGaps(sequence)
}
\arguments{
\item{sequence}{DNAString object instance}
}
\value{
character string of the sequence without gaps
}
\description{
Remove gaps (\'-\')
}
