% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignAnalyze.R
\name{alignAnalyze}
\alias{alignAnalyze}
\title{This will be THE function for alignment-analysis output}
\usage{
alignAnalyze(filename, orfName, annotated = T, outputDirectory,
  specNames, ...)
}
\arguments{
\item{filename}{alignment file name or directory name}

\item{orfName}{identifier for ORF. SGD name if annotated}

\item{annotated}{boolean shows if the given sequence is annotated or not}

\item{outputDirectory}{directory to save alignment output}

<<<<<<< HEAD
\item{specNames}{species names given by user provided tree. it is needed for correct output names and analysis}

\item{...}{placeholder for sequence file names}
}
\description{
This will be THE function for alignment-analysis output
=======
\item{...}{placeholder for sequence file names}
}
\description{
This will be the function for alignment-analysis output
>>>>>>> 148476c140b1b99b5faa0856ac47ec854c8eaa4f
}
