% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{This function reads all output files previously created and analyzes the outputs}
\usage{
analyze(orfName, path, vec)
}
\arguments{
\item{orfName}{orf identifier for the orf of interest}

\item{path}{folder path for output files}

\item{vec}{this is a vector to assess which sequences exists in this order scer,spar,smik,skud,sbay}
}
\description{
This function reads all output files previously created and analyzes the outputs
}
