% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{This function reads all output files previously created and analyzes the outputs}
\usage{
analyze(path, orfName, types)
}
\arguments{
\item{path}{folder path for output files}

\item{orfName}{orf identifier for the orf of interest}

\item{types}{this is a vector to assess which sequences exists}
}
\description{
This function reads all output files previously created and analyzes the outputs
}
