% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze msa
This function reads all output files previously created and analyzes the outputs}
\usage{
analyze(outputDirectory, orfName, types, best = T)
}
\arguments{
\item{outputDirectory}{folder path for reading input files}

\item{orfName}{orf identifier for the orf of interest}

\item{types}{this is a vector to assess which species sequences exists}

\item{best}{specifies whether to analyze only best overlapping orf or all.}
}
\description{
Analyze msa
This function reads all output files previously created and analyzes the outputs
}
