% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findORF.R
\name{findORF}
\alias{findORF}
\title{Find orf start and stop coordinates}
\usage{
findORF(aa_seq)
}
\arguments{
\item{AAString}{instance}
}
\value{
coordinates of M---X sequence if exists and longer than I---X, returns I---X if it exists and Longer than M--X
}
\description{
Find orf start and stop coordinates
}
