% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findYgeneSeq.R
\name{findYGeneSeq}
\alias{findYGeneSeq}
\title{Finds annotated ORF sequence from annotated sequence file downloaded from SGD}
\usage{
findYGeneSeq(geneName)
}
\arguments{
\item{orfName}{annotated orf Y----- name}
}
\value{
sequence of given orf name as DNAString instance
}
\description{
Finds annotated ORF sequence from annotated sequence file downloaded from SGD
}
