% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readFiles}
\alias{readFiles}
\title{Read syntetic block sequence files}
\usage{
readFiles(path, specNames)
}
\arguments{
\item{path}{folder path that has sequence fasta files in it}

\item{specNames}{species names to look for
every sequence file needs to contain its species name . (i.e. S. Cerevisiae file should contain 'scer')
also assumes there is only 1 file for every species}
}
\value{
DNAStringSet object with the sequences in it 
order of DNAStringSet will be order of specNames

TODO check for multiple files having the search string in them
}
\description{
Read syntetic block sequence files
}
